+++ lib/src/features/cipher/presentation/components/algorithm_selected.dart
@@ -1,11 +1,11 @@
 import 'package:cipher_dove/src/core/_core.dart';
-import 'package:cipher_dove/src/routing/app_router.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
 import '../../../../common_widgets/custom_button.dart';
 import '../../../../constants/_constants.dart';
+import '../../../../routing/app_router.dart';
 import '../../../../util/context_shortcut.dart';
 import '../cipher_mode_state.dart';
 
@@ -20,6 +20,7 @@ class AlgorithmSelected extends ConsumerWidget {
     return ConstrainedBox(
       constraints: BoxConstraints(maxWidth: kScreenWidth(context) * 0.55),
       child: CustomButton(
+        msg: "Select Algorithm",
         onTap: () {
           context.pushNamed(AppRoute.algorithmSelection.name);
         },