+++ lib/src/features/version_check/data/version_repo_.dart
@@ -1,0 +1,6 @@
+import '../domain/version_check.dart';
+
+abstract class VersionCheckRepo {
+  late final VersionCheck versionCheck;
+  Future<void> getVersionCheck();
+}