+++ lib/src/routing/app_router.dart
@@ -1,20 +1,24 @@
-import 'package:cipher_dove/src/features/cipher/presentation/algorithm_selection_screen.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
-import 'package:cipher_dove/src/common_widgets/hud_overlay.dart';
-import 'package:cipher_dove/src/core/app/navigation.dart';
-import 'package:cipher_dove/src/features/home/presentation/home_screen.dart';
-import 'package:cipher_dove/src/routing/not_found_screen.dart';
+
+import '../common_widgets/hud_overlay.dart';
+import '../core/_core.dart';
+import '../features/about/presentation/about_screen.dart';
+import '../features/cipher/presentation/algorithm_selection_screen.dart';
+import '../features/home/presentation/home_screen.dart';
+import 'not_found_screen.dart';
 
 part 'app_router.g.dart';
 
 enum AppRoute {
   home,
   algorithmSelection,
-  about;
+  about,
+  license,
+  unknown;
 
   const AppRoute();
 }
@@ -48,28 +52,41 @@ GoRouter goRouter(Ref ref) {
         builder: (context, state) {
           return const HudOverlay(child: AlgorithmSelectionScreen());
         },
-        routes: [],
         onExit: (context, state) {
           // Prevent keyboard suddenly opens when going back to home screen after
           // once focused on one of the textfields.
           FocusScope.of(context).unfocus();
           return true;
         },
+        routes: [],
       ),
       GoRoute(
-        path: AppRoute.about.path,
-        name: AppRoute.about.name,
-        pageBuilder: (context, state) {
-          return MaterialPage(
-              key: state.pageKey, // Recommended
-              fullscreenDialog: true,
-              child: Container(
-                width: 100,
-                height: 100,
-                color: Colors.black,
-              ));
+          path: AppRoute.about.path,
+          name: AppRoute.about.name,
+          builder: (context, state) {
+            return const HudOverlay(child: AboutScreen());
+          },
+          onExit: (context, state) {
+            // Prevent keyboard suddenly opens when going back to home screen after
+            // once focused on one of the textfields.
+            FocusScope.of(context).unfocus();
+            return true;
+          },
+          routes: [
+            GoRoute(
+              path: AppRoute.license.path,
+              name: AppRoute.license.name,
+              builder: (context, state) {
+                return const LicensePage();
+              },
+            ),
+          ]),
+      GoRoute(
+        path: "/404",
+        name: AppRoute.unknown.name,
+        builder: (context, state) {
+          return const NotFoundScreen();
         },
-        routes: [],
       ),
     ],
     errorBuilder: (context, state) {