+++ test/mocks.dart
@@ -1,10 +1,10 @@
 import 'dart:async';
 
+import 'package:cipher_dove/src/features/version_check/data/version_repo_.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:dio/dio.dart';
 import 'package:cipher_dove/src/constants/app_info.dart';
 import 'package:cipher_dove/src/core/network/api_service.dart';
-import 'package:cipher_dove/src/features/version_check/data/version_repo.dart';
 
 class MockDio extends Mock implements Dio {}